requirements.txt
astra-haystack==0.1.1
sentence-transformers==2.3.1
haystack-ai==2.0.0b5